pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "optimized-routing-extension"
version = "1.0.0"
description = "Optimized BlueFolder routing + Google Maps + Cloudflare shortener"
readme = "README.md"
authors = [
    { name = "David Durost" }
]
license = { text = "MIT" }
requires-python = ">=3.9"

dependencies = [
    "requests",
    "python-dotenv",
    "googlemaps",
    "pydantic",
    "tenacity",
]

[tool.setuptools.packages.find]
where = ["optimized_routing"]

[project.scripts]
optimized-routing = "optimized_routing.main:cli_entry"
